refactor: create datastore by factory method
This commit refactores datastore package in a way that it introduces
factory method so that the concrete instances can be created w/o the
need to modify main.go.

The concrete instance is now created based on the settings.Driver field.

Example:

	db, err := datastore.New(settings)
	err != nil {
		// DataStore not created at all
	}
